Add view link for a warning note in the records history
The creation of a warning note is visible in record history, but the ‘view’ link is missing.

This adds a condition to the case link component to render a `view` link to the warning note
within the records history.

It also adds a warning note id attribute to the case form data type so that it can be used to create the relevant path.