Only display end date / next review date on the most recent review
This adds a condition to check for the most recent review (which would be displayed first)
and pass a value for the end date / next review date if it is.

Since the end date and next review date are values saved on the original warning note,
we only want to pass these values in on the most recent review rather than being repeated for each review
on the page.