(Refactor) : Changes based on PR comments
- Uses a path parameter for the warning note view link:
This initially created a view link using query params
but it was suggested to change this to be a nested route
as there are primarily only three ways a user should see
the warning note details page depending on
the stage of the warning note i.e
* Warning note created
* Warning note reviewed
* Warning note ended

- It also refactors some of the code to improve readablity and
communicate the intention of each bit of functionality in a
(hopefully!) clearer way.

- This also adds a conditional render to the form steps type.
Before we were creating and editing a copy
and using that within the component to render the summary.
It was suggested that creating and manipulating copies like this could
get unwieldly and one would have to remember all the different "copies" of the code
and why they exist in the niches
 which could make the code harder to maintain in the future.