Undo change to reviewNotes component in review form step
This was initially changed because the reviewNotes did not match the property
in the API response when we make a GET request for warning note reviews.

The API response was returning a Notes property for each warning note review object.

From discussing the current naming conventions for the warning note data model,
it was agreed that any details of a review should be saved as `review notes`
and `notes` should be reserved for the original warning note.

We've made a change in the API to reflect this, so this reverses the previous change
made in the frontend.