Add missing "s" in URL path (#446)
When we click on `save and finish later`
we are redirected to a page not found. URL path: `/form-in-progress`

We checked the URL of the forms to complete tab in the dashboard
and realised that the path for the forms tab
was: `/forms-in-progress`

This corrects the path when clicking the save and finish later button
to redirect the user to the dashboard forms to complete tab.