Only add `DisclosedWithIndividual` to the PATCH request body when defined
We don't capture a value of `DisclosedWithIndividual` in the warning note review form for a child.
We want this value to remain null for a child.

Before we create the PATCH request for an adult warning note review,
we convert the selected option for `DisclosedWithIndividual` to the appropriate boolean,
however, because of the condition we use to pass the appropriate boolean,
if the `DisclosedWithIndividual` was undefined, it would return false.

This checks that the `DisclosedWithIndividual` value is NOT undefined,
before converting the string option 'Yes'/'No' into true/false and adding the parameter to the PATCH request.