feat(SCT-481): add Cypress test for relationships
This goes through the flow of adding a relationship, viewing it and
then removing it.

In the test, we had to intercept the request for creating a relationship
because the cookie used by Cypress is associated with a worker that does
not exist. A worker with `e2e.tests.adult@hackney.gov.uk` has been added
so that a valid email address for `createdBy` could be used.