Fix: Address mismatch between creating a new note and a person's details
When a person's address is updated,
 the change is reflect in the UI when viewing the person details page.

However, when we make address changes, the previous addresses still exist
on the record as a list of addresses.

When creating a new case note or submission, the widget was pulling the first address
from this list of addresses which meant it was not matching the address shown in the person's details.

This changes the code that retrieves the address we want to use when a person has a list of addresses
associated with them and gets the last address added instead of the first.

We also added data_test_id attributes to make testing the change easier.