Creates new form_fields array to avoid duplication
There was a UI bug that occured when a user goes back to the add page after viewing the review page.

When they go back, the CP category selection would get repeated.

This creates a new array for the form field everytime the component is rendered, which stops the duplication bug