SCT-1394 Add prefiled field value on start date when canceled LAC review
When we go from the Update edit page to the Review Update edit page we carry the start date value with us to be shown on the Review page.
However if we click cancel and return to the Update edit page we loose the values for the validation on setting the new start date.
Ie: currentCaseStatusStartDate is undefined