Fix Boolean value for isScheduledCaseStatus
The Announcement component was being rendered to each update and edit page as we were getting an incorrect value for isScheduledCaseStatus.
Casting the value from a query String to a Number and then to a Boolean value, on the page that the value was needed as a Boolean, should keep it working correctly.