feat(SCT-1030): Mash dashboard  (#734)
* WIP MASH dashboard

code for start of the team assignments page, with static tabs

* Code to progress team assignments tab

Code to make the tabs bold as you switch between them

* fetch all mash referrals, filter by stage

* Integrating the WIP MASH cards

Moving the dummy data into the MASH cards

* Adjustments to the MASH card component and removing link on dashboard to team assignments

Edited the code for the MASH cards so they look closer to the designs. Also removed the link to team-assignments on the dashboard page.

* create factory builder for mash referral testing

* test api calling for fethcing all mash referrals

* update import path

* test information renders to MashCard

* PR based changes

removing unused CSS and making things such as 'rem' and 'px' consistent etc.

* update commit

* update commit

* Move team assignments code into a component

Created a component called 'MashDashboard' to move the team assignments code to so it is easier to test. Alongside this I made adjustments to the 'MashCard' code so it includes an Action button.

* Creating cards based on each tab/stage of MASH

Split the cards that show referral information into 4 cards per stage.

* Remove comments

Remove unwanted comments

* Pushing changes as per the PR comments

removing in text css, correcting some of the types, started changing <a/> tags to <Link/> components

* Reorganising 'teamassignment.tsx' and decision cards

Passed down the MashReferral array to each of the card types instead of individual parts of the MashReferral array.

* Changed <a/> to <Link/> components

Changed <a/> tags in the Mash Cards to <Link/> components

* Adjust href links, reorganise 'maincard.tsx' and correct client name

Adjust href links so they are more representative, reorganise 'maincard.tsx' so the code is less repetitive and correct client name so it includes a number if there are multiple clients.

* tidy up props/names

* update mash referral with screening types

* update mash referral factory with new screening properties

Co-authored-by: JohnFarrellDev <johnfarrell@fastmail.com>