feat(SCT-1030): WIP MASH action pages (#728)
* (WIP) FE for action pages

Started the FE code for the initial decision, screening and final decision pages.

* Import gov components

Utilising radio and selector components in screening.tsx

* WIP - MASH heading component

Created a component for the top of MASH action pages

* MASH heading component

altered the scss

* WIP - steps component

Coded a steps component will try make the next version more reusable so it can be used on the other action pages

* Making 'numbered steps' a reusable pattern

Turned the numbered steps into a reusable pattern so it can be used on other action pages.

* Small adjustments to 'NumberedSteps' css

Made slight adjustments to NumberedSteps.module.scss so things align slightly better

* FE for initial decision and final decision

corrected naming conventions, applied the 'number steps' to initial decision and final decision pages. Adjusted the css for part of the heading.

* WIP - MASH cards/boxes for team assignments

Working on coding a component to be reused for the different tabs of team assignments.

* Clean up code

Removed unused MASH card code so when it comes to merging there aren't conflicts. Also removed unused CSS.

* (WIP) Storing the options selected as states

Storing the options selected in the screening decision and whether a case is urgent as states.

* WIP - preset the selector options

Setting the selector so option 1 is preset and the radio buttons act as expected.

* move pages for mash referral decisions into path that includes referral id

* get referral by id

* for screening decision get the referral data, if wrong type or doesnt exist redirect to 404

* on form submit call submit screening decision

* implement API call for submitting a screening decision

* remove unrequired react fragment

* remove important usage from CSS

* move back work-in-progress page

* update request to submit worker email when patching referral

Co-authored-by: John Farrell <johnfarrell@fastmail.com>