SCT-1378 Add Edit announcement for changing start date consequences
If changing the start date of the current case status, this will affect the end date of the previous case status. This announcement is to make the user more aware of this.

Separate but related, i added an E2E test for displaying an announcement on the LAC Update page when making a scheduled case status and there is a pre-existing case status