feat(SCT-1446): Tests for the MASH cards and Headings (#753)
* WIP Mash Card tests

Pushing my progress on the WIP Mash Card tests

* Test for MashHeading

Pushing the test for the 'MashHeading/Heading.tsx'

* Test for MainCard.tsx

Created test for MainCard.tsx renders the 4 types of cards and checks the right components are rendered based on the filter type.

* PR corrections

Made some corrections based on the PR e.g. correcting the props we are testing for InitialDecisionCard.tsx and testing that the full text is tested '4 hours left' is tested on MashHeading.tsx

* Test amendments

Adjusting the test e.g. removing 'referral category' as an input to be expected for the initial decision card and changing the string to be expected so it isn't multiple elements.