feat(SCT-1437): Making time tags dynamic (#762)
* WIP time tag code

Pushing WIP code for time deadlines per the RAG ratings system and also the reverse timer for the business support tab.

* Adding time tag component to action pages

Swapped the hardcoded time tags in the action pages to the more dynamic time tags.

* Making time received dynamic

Changed time received to be dynamic by passing 'mashReferral.createdAt' to the component.

* Making time thresholds dynamic

Using 'mashReferral.createdAt' rather than the dummy time left.

* Making the thresholds dependent on the initial decision choice

Making the time thresholds based on the choices selected in initial decision

* Update MashTags.tsx

* Resolve conflicts and pass props to MashHeading

Resolved merge conflicts, and passed props to the Mash Heading 'components/MashHeading/Heading.tsx' since the time tag is placed within the Heading component.

* Adjust the react tests

Adjusted the react tests to include the filter props

* Changes based on PR

- reducing the props passed through
- altering the tests
- changing '==' to '==='
- when checking things based on decisions using 'toUpperCase()' to reduce risk of errors

* Add props

Passing mashReferrals to FinalDecision.

* First part of the MashTag tests

Pushing the first test for the contact stage reverse timer.

* Updates to tests script for MASH tags

Updated the test scripts for different conditions and stages. Test scenario needs to be done for overdue.

* Adding tests for the overdue conditions

Created tests for when screening and final decisions are overdue for each rag rating.

* Change code based on requirements and rename variables

Adjusted based on meeting with MASH team. Also adjusted the tests based on this.

* Refactoring MASH tags

The code has been refactored to reduce repetition.

* Adding test for 'mins left'

Creating a test for 'minutes left' for initial decision stage.

* Added the 'mins left' functionality when time left is less than an hour

Added code so that the mins left is rendered when the time left is in minutes when time left is less than an hour. Alongside the respective tests.

* Pushing form Prop changes

Pushing changes due to unit test issue.

* Remove unneccessary condition

Remove 'timeLeftinMilliseconds / 60 / 1000 > 0' as a condition since it is already checking if 'Overdue=== false'.

* Adjustments for PR

Changing '==' to '==='. Alongside renaming the 'output' function so it is more informative.

* Removing the 'moments' from the code

Altering the code to remove the 'moments'.

* Rounding the hours on the time tags and correcting the tests

Adding the 'Math.round()' to the hours left and adjusting the tests for the referral cards

Co-authored-by: John Farrell <johnfarrell@fastmail.com>