feat(SCT-1438): MASH high priority banner (#780)
* CSS for high priority banner

Pushing the CSS for the WIP high priority banner. the banner will show when the referral is high priority, through conditional classes.

* Add 'meter'/shadowing back

Adding the meter/shadow back into the MashCard CSS.

* Altered tests to include a test for the high priority banner

Edited the tests to include a tests for checking whether the high priority banner appears.

* Removed ternary operator and unnecessary {' '}

Swapped ternary operator to remove an empty string and change to '&&'. Also removed an unnecessary {' '}.

* WIP Priority banner component

Pushing the high priority banner as a reusable component.

* implement priority banner component into all 4 cards

* test the priority banner

* delete main card unused component

* correct deleted file

* Updating the text in tests

Updated some of the test descriptions so they are more appropriate.

Co-authored-by: JohnFarrellDev <johnfarrell@fastmail.com>