Fix interception in form Cypress test
The url property expects a full request path i.e. including http://hostname,
we just are specifying the path here so we should use the path property
instead.

Also, as of version 7 of Cypress intercept has become stricter so doesn't
match a substring so we need the "/api" before the "/test".